+++ src/material/snack-bar/testing/BUILD.bazel
@@ -20,29 +20,13 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/overlay",
         "//src/cdk/testing",
         "//src/cdk/testing/private",
-        "//src/cdk/testing/testbed",
-        "//src/material/snack-bar",
-        "@npm//@angular/platform-browser",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/cdk/testing",
         "//src/cdk/testing/testbed",
         "//src/material/snack-bar",
         "@npm//@angular/platform-browser",
